+++ discord/ui/core.py
@@ -27,6 +27,7 @@ from __future__ import annotations
 import asyncio
 import time
 from itertools import groupby
+from operator import attrgetter
 from typing import TYPE_CHECKING, Any, Callable
 
 from ..utils import find, get
@@ -41,10 +42,35 @@ if TYPE_CHECKING:
     from .view import View
 
 
+def _item_getter(iterable, **attrs) -> Item | None:
+    _all = all
+    attrget = attrgetter
+    for i in iterable:
+        converted = [
+            (attrget(attr.replace("__", ".")), value) for attr, value in attrs.items()
+        ]
+        try:
+            if _all(pred(i) == value for pred, value in converted):
+                return i
+        except AttributeError:
+            pass
+        if hasattr(i, "get_item"):
+            if child := i.get_item(**attrs):
+                return child
+    return None
+
+
 class ItemInterface:
     """The base structure for classes that contain :class:`~discord.ui.Item`.
 
     .. versionadded:: 2.7
+
+    .. container:: operations
+
+        .. describe:: len(x)
+
+            Returns the total count of all items in this interface.
+            This includes items that contain other items, which count towards Discord's component limits.
 
     Parameters
     ----------
@@ -53,6 +79,8 @@ class ItemInterface:
     timeout: Optional[:class:`float`]
         Timeout in seconds from last interaction with the UI before no longer accepting input. Defaults to 180.0.
         If ``None`` then there is no timeout.
+    store: Optional[:class:`bool`]
+        Whether this interface should be stored for callback listening. Setting it to ``False`` will ignore callbacks and prevent item values from being refreshed. Defaults to ``True``.
 
     Attributes
     ----------
@@ -61,8 +89,6 @@ class ItemInterface:
         If ``None`` then there is no timeout.
     children: List[:class:`Item`]
         The list of children attached to this structure.
-    store: Optional[:class:`bool`]
-        Whether this interface should be stored for callback listening. Setting it to ``False`` will ignore callbacks and prevent item values from being refreshed. Defaults to ``True``.
     """
 
     def __init__(
@@ -85,6 +111,12 @@ class ItemInterface:
 
     def __repr__(self) -> str:
         return f"<{self.__class__.__name__} timeout={self.timeout} children={len(self.children)}>"
+
+    def __len__(self) -> int:
+        return sum(len(i) for i in self.children)
+
+    def __bool__(self) -> bool:
+        return True
 
     async def _timeout_task_impl(self) -> None:
         while True:
@@ -115,30 +147,54 @@ class ItemInterface:
     def to_components(self) -> list[dict[str, Any]]:
         return [item.to_component_dict() for item in self.children]
 
-    def get_item(self, custom_id: str | int) -> Item | None:
-        """Gets an item from this structure. Roughly equal to `utils.get(self.children, ...)`.
+    def get_item(self, custom_id: str | int | None = None, **attrs: Any) -> Item | None:
+        r"""Gets an item from this structure. Roughly equal to `utils.get(self.children, **attrs)`.
         If an :class:`int` is provided, the item will be retrieved by ``id``, otherwise by  ``custom_id``.
         This method will also search nested items.
+        If ``attrs`` are provided, it will check them by logical AND as done in :func:`~utils.get`.
+        To have a nested attribute search (i.e. search by ``x.y``) then pass in ``x__y`` as the keyword argument.
+
+        Examples
+        ---------
+
+        Basic usage:
+
+        .. code-block:: python3
+
+            container = my_view.get(1234)
+
+        Attribute matching:
+
+        .. code-block:: python3
+
+            button = my_view.get(label='Click me!', style=discord.ButtonStyle.danger)
 
         Parameters
         ----------
-        custom_id: Union[:class:`str`, :class:`int`]
+        custom_id: Optional[Union[:class:`str`, :class:`int`]]
             The id of the item to get
+        \*\*attrs
+            Keyword arguments that denote attributes to search with.
 
         Returns
         -------
         Optional[:class:`Item`]
-            The item with the matching ``custom_id`` or ``id`` if it exists.
+            The item with the matching ``custom_id``, ``id``, or ``attrs`` if it exists.
         """
-        if not custom_id:
-            return None
-        attr = "id" if isinstance(custom_id, int) else "custom_id"
-        child = find(lambda i: getattr(i, attr, None) == custom_id, self.children)
-        if not child:
-            for i in self.children:
-                if hasattr(i, "get_item"):
-                    if child := i.get_item(custom_id):
-                        return child
+        child = None
+        if custom_id:
+            attr = "id" if isinstance(custom_id, int) else "custom_id"
+            if attrs:
+                attrs[attr] = custom_id
+            child = find(lambda i: getattr(i, attr, None) == custom_id, self.children)
+            if not child:
+                for i in self.children:
+                    if hasattr(i, "get_item"):
+                        if child := i.get_item(custom_id):
+                            return child
+        elif attrs:
+            child = _item_getter(self.children, **attrs)
+
         return child
 
     def add_item(self, item: Item) -> Self: